requirements.txt
asgiref==3.11.0
crispy-bootstrap5==0.7
Django==4.2.7
django-crispy-forms==2.0
django-filter==23.3
django-widget-tweaks==1.5.0
gunicorn==23.0.0
packaging==25.0
Pillow==10.0.0
sqlparse==0.5.4
tzdata==2025.2
whitenoise==6.11.0
